zap/requirements.txt
aiohttp==3.9.0
codedx-api @ git+https://github.com/broadinstitute/dsp-appsec-codedx-api-client-python.git@cf0de6fcdd0a4db98f2c2a35bd9d2ae7e8056d6b
google-auth==1.30.0
google-cloud-pubsub==2.5.0
google-cloud-storage==1.38.0
requests==2.31.0
slack-sdk==3.5.1
defusedxml==0.7.1
pytz
zaproxy==0.2.0
google-api-python-client==2.116.0